+++ source/extension/unity/Package/Runtime/Compatibility/IsExternalInit.cs
@@ -1,0 +1,8 @@
+#if UNITY_2021_2_OR_NEWER
+namespace System.Runtime.CompilerServices
+{
+internal static class IsExternalInit
+{
+}
+}
+#endif